+++ R/outcome-helpers.R
@@ -145,8 +145,8 @@
 .outcome_data_likelihood_sei <- function(object) {
 
   if (.is_data_known_v(object[["data"]])) {
-    if ((.is_data_known_v_parameterization(object[["data"]], "whitened") ||
-         .is_data_known_v_parameterization(object[["data"]], "block_mvn")) &&
+    if ((.is_data_known_v_backend(object[["data"]], "whitened") ||
+         .is_data_known_v_backend(object[["data"]], "block_mvn")) &&
         .data_known_v_correlated(object[["data"]])) {
       stop(
         "Original-scale pointwise likelihood standard errors are not available ",